Patch, Put, Delete
